Assignment 4
This assignment is about industry talk with Ms. Qistina Batrisyia. She was also an UTM student before and her experience sharing is very useful to me. From her talk, I learnt many about how I need to be get prepare if I want to take IT career in the future. She also sharing many tools and the function each of them to us during the talk. I think that it is very helpful. As a conculsion, this assignment is a very good chance to me to get in touch with knowledge of IT career and also the experience from our senior. I hope that there are more activities ar event like this being held in the future.